+++ bedspread-core/src/test/java/it/cnr/iasi/leks/bedspread/tests/BasicSemanticSpreadOnDBpediaTest.java
@@ -35,7 +35,7 @@ import it.cnr.iasi.leks.bedspread.rdf.KnowledgeBase;
 import it.cnr.iasi.leks.bedspread.rdf.impl.RDFFactory;
 import it.cnr.iasi.leks.bedspread.rdf.impl.RDFGraph;
 import it.cnr.iasi.leks.bedspread.rdf.sparqlImpl.DBpediaKB;
-import it.cnr.iasi.leks.debspread.tests.util.PropertyUtilNoSingleton;
+import it.cnr.iasi.leks.bedspread.tests.util.PropertyUtilNoSingleton;
 
 import org.junit.Assert;
 import org.junit.Test;
