+++ bedspread-core/src/test/java/it/cnr/iasi/leks/bedspread/tests/PolicentricSemanticSpreadTest.java
@@ -39,8 +39,8 @@ import it.cnr.iasi.leks.bedspread.rdf.AnyResource;
 import it.cnr.iasi.leks.bedspread.rdf.KnowledgeBase;
 import it.cnr.iasi.leks.bedspread.rdf.impl.RDFFactory;
 import it.cnr.iasi.leks.bedspread.rdf.impl.RDFGraph;
+import it.cnr.iasi.leks.bedspread.tests.util.PropertyUtilNoSingleton;
 import it.cnr.iasi.leks.bedspread.util.SetOfNodesFactory;
-import it.cnr.iasi.leks.debspread.tests.util.PropertyUtilNoSingleton;
 
 public class PolicentricSemanticSpreadTest extends AbstractTest{
