+++ bedspread-core/src/test/java/it/cnr/iasi/leks/bedspread/tests/util/PropertyUtilNoSingleton.java
@@ -16,7 +16,7 @@
  *	 You should have received a copy of the GNU General Public License
  *	 along with this source.  If not, see <http://www.gnu.org/licenses/>.
  */
-package it.cnr.iasi.leks.debspread.tests.util;
+package it.cnr.iasi.leks.bedspread.tests.util;
 
 import it.cnr.iasi.leks.bedspread.config.PropertyUtil;
